Soft physics and organization Made my first bit of soft body physics ever. YAY! I create a bunch of distance constraints between the pointmasses that act like springs. They pull the pointmasses together if they get too far and push them away if they get too close. This way object will kinda keep its shape. But it's not really perfect yet. I modify the pointmasses data during the constraint calculations which I think can cause some issues, but I haven't yet tested it. Also I for some reason don't get as many constraints as I think I should get, resulting in objects not really keeping their shape at lower stiffness values. They are either very rigid or don't keep their shape really at all. Whether this difference is because I'm stupid and miscalculated the amount of constraints I should have or because I'm stupid and made a mistake in the code.
Also, this fucks up the lighting because these distance constraints
cause the object to rotate a decent bit, which definitely is a rather
challenging bug to fix, but I don't yet account for the rotation in my
vertex normals so the normals just stay pointing in the same direction
no matter how hard the object spins around. Gotta love normals, when I
encounter a severe lighting problem, 90% of the time it has something to
do with fucked up normals.

And now to organization. I moved allmost all of my stuff to different
functions and grouped them in 2 separate files, one for simulation and
one for rendering. Much less messy now.