Beginnings of a tetralizer I have now begun my quite likely painful journey to creation my own triangle mesh tetralizer. Tetralized mesh should be easier to do soft body physics on. So far I added a way to render the simulation mesh as a wireframe, because from here on the simulation mesh may not be the same as the rendered mesh. I also completed the very first step of tetralization, which is calculating a "super tetrahedron", a tetrahedron containing all the vertices of the triangle mesh. I also of course had to adapt the code figuring out all the pointmasses and constraints too.
It's not perfect as there certainly will be duplicate constraints and
propably pointmasses too, but that doesn't matter yet as I still only
have 1 tetrahedron. Also doing physics on the tetrahedralized mesh is
gonna give some wacky results as the rendered mesh is not properly mapped
to the simulation mesh yet, but the physics themself should work
okay-ish.