Speed and low quality I started the work on improving the speed and memory usage of the tetrahedrilizer so that I could tetrahedrilize more complex meshes. What I did to improve the speed is switch from storing the coordinates of the tetrahedron to storing the index to the array of coordinates. This reduces memory usage, but more importantly speeds up the proces of getting rid of duplicate data, which there is a lot of. I also noticed that the initial tetrahedron generator has a habbit of creating exponentially more duplicate tetrahedrons. It first creates a couple and those each create a couple more and so on until the basic cube, which I think should only need 24 tetrahedrons, had well over 2000. And this problem is certainly exponentially worse for more complex models. Using the help of storing stuff using indeces I am able to quickly and easily deduplicate the data every iteration of the tetrahedron generation loop, gettin rid of this exponential waste of resources. Not only does this dramatically reduce memory usage and speed up tetrahedron generation, but it also makes later steps faster too.
And this change clearly does work. Previously tetralizing the basic cube
took over 80 milliseconds, now it takes about 0,2 milliseconds. I
haven't measured the memory usage, but it too surely is much lower. The
issue however is that it doesn't quite give correct results. It has a
couple of tetrahedrons missing and most of them are somehow connected to
the vertices of the original super tetrahedron. That is bad, cause those
vertices aren't part of the model and thus I remove them at the end to
get rid of unneeded tetrahedrons. But because most tetrahedrons are for
some reason connected to the super tetrahedron, deleting super
tetrahedron ends up removing lots of needed tetrahedrons too.
Fixing those quality issues will increase the tetrahedron count and thus
surely the tetralizing time too, but even if the time increases 10x,
this is still gonna be a 40x performance improvement, so it's a good
start.