A lot of work for no gains at all I decided to add the facet recory system, which as you can see from the lines of code added, was a decent bit of work for a single commit. But I managed to get it done. Well, "done". It's technically working, but isn't really producing any good results. The reason for this being it took me a while to get here and so I couldn't be bothered to try to hunt for bugs in this massive mess of a code right now. So you'll just have to deal with badly functioning code. Btw, the this code isn't functional is in it deletes useful tetrahedrons without preventing other useful tetrahedrons from being deleted, which is it's main job.
But in the name of honesty, I must admit that not all of this was really
written in this session. The gift wrapping code, which is about 60 lines
of code, was taken from one of my previous commits. But everything else
was written by me within the last few hours of making this commit.